# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/oauth/google": {
      "post": {
        "description": "Retorna URL de OAuth de Google. Frontend debe redirigir al usuario a esta URL.",
        "operationId": "OAuthController_googleSignIn_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthInitiateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL de OAuth generada",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://accounts.google.com/o/oauth2/v2/auth?client_id=..."
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al iniciar OAuth"
          }
        },
        "summary": "Iniciar autenticación con Google",
        "tags": [
          "OAuth"
        ]
      }
    },
    "/v1/auth/oauth/apple": {
      "post": {
        "description": "Retorna URL de OAuth de Apple. Frontend debe redirigir al usuario a esta URL.",
        "operationId": "OAuthController_appleSignIn_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthInitiateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL de OAuth generada",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://appleid.apple.com/auth/authorize?client_id=..."
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al iniciar OAuth"
          }
        },
        "summary": "Iniciar autenticación con Apple",
        "tags": [
          "OAuth"
        ]
      }
    },
    "/v1/auth/oauth/callback": {
      "get": {
        "description": "Procesa el callback de Google/Apple después de autenticación exitosa. Auto-crea usuario si es primera vez. Retorna accessToken y datos de usuario.",
        "operationId": "OAuthController_handleCallback_v1",
        "parameters": [
          {
            "name": "access_token",
            "required": true,
            "in": "query",
            "description": "Access token de Supabase",
            "schema": {
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
              "type": "string"
            }
          },
          {
            "name": "refresh_token",
            "required": false,
            "in": "query",
            "description": "Refresh token de Supabase",
            "schema": {
              "example": "v1.abc123...",
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "description": "Provider usado (google, apple)",
            "schema": {
              "example": "google",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJhbGc...",
                    "refreshToken": "v1.abc...",
                    "user": {
                      "id": "uuid-123",
                      "email": "user@gmail.com",
                      "name": "Juan",
                      "paternal_last_name": "Pérez",
                      "maternal_last_name": "González",
                      "avatar": "https://lh3.googleusercontent.com/...",
                      "google_connected": true,
                      "apple_connected": false
                    },
                    "needsPostRegistration": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token de OAuth inválido"
          }
        },
        "summary": "Manejar callback de OAuth",
        "tags": [
          "OAuth"
        ]
      }
    },
    "/v1/auth/oauth/providers": {
      "get": {
        "description": "Retorna los providers de OAuth (Google, Apple, Facebook) conectados al usuario.",
        "operationId": "OAuthController_getConnectedProviders_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Providers conectados",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "google_connected": true,
                    "apple_connected": false,
                    "fb_connected": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener providers conectados",
        "tags": [
          "OAuth"
        ]
      }
    },
    "/v1/auth/oauth/{provider}": {
      "delete": {
        "description": "Desconecta un provider de OAuth (google, apple, fb) del usuario. Nota: Esto solo actualiza el flag en BD, no desvincula la cuenta en Supabase.",
        "operationId": "OAuthController_disconnectProvider_v1",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "Provider a desconectar",
            "schema": {
              "enum": [
                "google",
                "apple",
                "fb"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider desconectado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Provider desconectado exitosamente"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Provider inválido"
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desconectar un provider",
        "tags": [
          "OAuth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OAuthInitiateDto": {
        "type": "object",
        "properties": {
          "redirectUrl": {
            "type": "string",
            "description": "URL de redirección después de completar OAuth",
            "example": "https://sacdia.app/auth/callback"
          }
        }
      },
...
    }
  }
}